#2: provisionResource works

+++ bowler-kernel/hardware/src/test/kotlin/com/neuronrobotics/bowlerkernel/hardware/registry/RegistryTestUtil.kt
@@ -72,7 +72,7 @@ internal fun HardwareRegistry.makeDeviceOrFail(id: String): MockDevice =
 
 internal fun HardwareRegistry.makeDeviceResourceOrFail(
     device: Device,
-    attachmentPoint: Byte
+    attachmentPoint: Int
 ): MockUnprovisionedDeviceResource =
     registerDeviceResource(
         device,
